+++ src/app/about/reakit/index.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+class Reakit extends React.Component {
+    render() {
+        return(
+            <p>The id is: {this.props.match.params.about}</p>
+        );
+    }
+}
+
+export default Reakit;