+++ changelog.md
@@ -1,2 +1,4 @@
 + [FEA] : Scaffold base for Reakit usage
 + [FEA] : Script to automate project
++ [ENH] : Include `react-router-dom` to manage routes
++ [BUG] : Include app folder